[REFACTOR] Changed position of Category and CategoryWithID
