+++ Eventy/app/src/main/java/com/example/eventy/solutions/model/CategoryWithID.java
@@ -1,4 +1,4 @@
-package com.example.eventy.model.solution;
+package com.example.eventy.solutions.model;
 
 import com.example.eventy.model.enums.Status;
